fix(frontend): resolve interface bugs — TypeScript errors, Tailwind v4, missing proxy, and integrate IPAColorDisplay
- Fix 7 TypeScript errors in LearnPage.tsx by adding explicit type annotations
  for drill targets/pairs arrays with optional audio properties
- Fix Tailwind v4 setup: add @import "tailwindcss" to theme.css and remove
  incompatible v3 tailwind.config.cjs (v4 uses CSS-based config)
- Add missing /api proxy route in vite.config.ts for learn/lesson endpoints
- Consolidate duplicate types: types/index.ts now re-exports from types/api.ts
  instead of maintaining divergent copies (api.ts has display field, index.ts didn't)
- Add HealthResponse/ComponentHealth to api.ts canonical types
- Export and integrate IPAColorDisplay component into PracticePage results
  (was built but never wired up)
