+++ frontend/src/components/results/index.ts
@@ -2,3 +2,4 @@ export { ScoreRing } from './ScoreRing';
 export { AlignmentTable } from './AlignmentTable';
 export { InlineTokens } from './InlineTokens';
 export { ErrorSummary } from './ErrorSummary';
+export { IPAColorDisplay } from './IPAColorDisplay';
