+++ frontend/src/theme.css
@@ -1,3 +1,5 @@
+@import "tailwindcss";
+
 /* ==============================================
  * PronunciaPA - Unified Design System
  * Dark-first Glassmorphism + Light toggle
