--- frontend/tailwind.config.cjs
@@ -1,10 +1,0 @@
-module.exports = {
-  content: [
-    './public/**/*.html',
-    './src/**/*.{js,ts,jsx,tsx}'
-  ],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-};
