+++ frontend/vite.config.ts
@@ -15,6 +15,7 @@ export default defineConfig({
     port: 5173,
     proxy: {
       '/v1': 'http://localhost:8000',
+      '/api': 'http://localhost:8000',
       '/health': 'http://localhost:8000',
       '/ws': {
         target: 'ws://localhost:8000',
